Chemical Viscosity Prediction Model
The final project for my Capstone in Machine Learning as part of my Master's in Data Science degree from Eastern University. For the project, I created a machine learning model to predict the resultant viscosity of a chemical used to line the interior of aluminum cans in a production environment. The QA manager at the plant hypothesized that the weather may have an impact on the process, so I incorporated historical weather data from the nearest monitoring facility to the plant into the dataset. 

This is a full machine learning proejct encompassing data cleaning, preprocessing, feature selection, model selection, model tuning, and creating a user-interface Flask.